rename FloatingChat to DirectMessages
