--- src/components/@shared/Orbis/Message.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function Message() {
-  return <div>Message</div>
-}
